refactor: extract aggregation module from store and actor
Isolate committee-signature aggregation code in its own module so the
async worker orchestration, job types, and pure aggregation helpers live
together. The store now focuses on attestation/block state, and the
actor (lib.rs) only handles session lifecycle hooks on BlockChainServer.

No behavior change.
